PPLUS-510: Automatically apply code changes (with integrator) during an upgrade / tests fix
