+++ composer.json
@@ -52,7 +52,8 @@
     "config": {
         "sort-packages": true,
         "allow-plugins": {
-            "dealerdirect/phpcodesniffer-composer-installer": true
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "spryker-sdk/integrator": false
         }
     },
     "scripts": {
