+++ tests/CodeComplianceTest/Domain/Checks/Filters/BaseFilterTest.php
