+++ tests/CodeComplianceTest/Domain/Checks/Filters/IgnoreListFilterTest.php
@@ -5,14 +5,12 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace CodeComplianceTest\Domain\Filters;
+namespace CodeComplianceTest\Domain\Checks\Filters;
 
 use Codebase\Application\Dto\ClassCodebaseDto;
-use CodeCompliance\Domain\Checks\Filters\PluginFilter;
-use CodeComplianceTest\Domain\Checks\Filters\BaseFilterTest;
-use ReflectionClass;
+use CodeCompliance\Domain\Checks\Filters\IgnoreListFilter;
 
-class PluginFilterTest extends BaseFilterTest
+class IgnoreListFilterTest extends BaseFilterTest
 {
     /**
      * @return array<string>
@@ -20,8 +18,12 @@ class PluginFilterTest extends BaseFilterTest
     public function provideClassNamesData(): array
     {
         return [
-            'TestProject\Method\ProjectPlugin',
-            'TestCore\Method\CoreMethod',
+            'CoreTest/Test/FooClass',
+            'CoreTest/Kernel/TestClass',
+            'CoreTest/Development/TestClass',
+            'CoreTest/Test/TestBootstrap',
+            'CoreTest/Test/TestConfigurationProvider',
+            'Spryker/Zed/TestDataImport',
         ];
     }
 
@@ -31,7 +33,7 @@ class PluginFilterTest extends BaseFilterTest
     public function testFilter(): void
     {
         // Arrange
-        $businessFactoryFilter = new PluginFilter();
+        $businessFactoryFilter = new IgnoreListFilter();
 
         // Act
         $filteredSources = $businessFactoryFilter->filter($this->getCodebaseObjects());
@@ -48,9 +50,12 @@ class PluginFilterTest extends BaseFilterTest
      */
     protected function createClassCodebaseDtoWithClassName(string $className): ClassCodebaseDto
     {
+        $parent = new ClassCodebaseDto(['Core']);
+        $parent->setClassName($className);
+
         $dto = new ClassCodebaseDto(['Core']);
-        $dto->setClassName($className);
-        $dto->setReflection(new ReflectionClass($className));
+        $dto->setClassName('Project/Foo/Class');
+        $dto->setParent($parent);
 
         return $dto;
     }
