+++ tests/CodeComplianceTest/Domain/Checks/Filters/PluginFilterTest.php
@@ -5,13 +5,13 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace CodeComplianceTest\Domain\Filters;
+namespace CodeComplianceTest\Domain\Checks\Filters;
 
 use Codebase\Application\Dto\ClassCodebaseDto;
-use CodeCompliance\Domain\Checks\Filters\IgnoreListFilter;
-use CodeComplianceTest\Domain\Checks\Filters\BaseFilterTest;
+use CodeCompliance\Domain\Checks\Filters\PluginFilter;
+use ReflectionClass;
 
-class IgnoreListFilterTest extends BaseFilterTest
+class PluginFilterTest extends BaseFilterTest
 {
     /**
      * @return array<string>
@@ -19,12 +19,8 @@ class IgnoreListFilterTest extends BaseFilterTest
     public function provideClassNamesData(): array
     {
         return [
-            'CoreTest/Test/FooClass',
-            'CoreTest/Kernel/TestClass',
-            'CoreTest/Development/TestClass',
-            'CoreTest/Test/TestBootstrap',
-            'CoreTest/Test/TestConfigurationProvider',
-            'Spryker/Zed/TestDataImport',
+            'TestProject\Method\ProjectPlugin',
+            'TestCore\Method\CoreMethod',
         ];
     }
 
@@ -34,7 +30,7 @@ class IgnoreListFilterTest extends BaseFilterTest
     public function testFilter(): void
     {
         // Arrange
-        $businessFactoryFilter = new IgnoreListFilter();
+        $businessFactoryFilter = new PluginFilter();
 
         // Act
         $filteredSources = $businessFactoryFilter->filter($this->getCodebaseObjects());
@@ -51,12 +47,9 @@ class IgnoreListFilterTest extends BaseFilterTest
      */
     protected function createClassCodebaseDtoWithClassName(string $className): ClassCodebaseDto
     {
-        $parent = new ClassCodebaseDto(['Core']);
-        $parent->setClassName($className);
-
         $dto = new ClassCodebaseDto(['Core']);
-        $dto->setClassName('Project/Foo/Class');
-        $dto->setParent($parent);
+        $dto->setClassName($className);
+        $dto->setReflection(new ReflectionClass($className));
 
         return $dto;
     }
